ml-server/requirements.txt
Flask>=2.0
gunicorn>=20.0
ultralytics
Pillow
torch
torchvision
# Add other dependencies your model needs, e.g., numpy, opencv-python-headless
